% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Production_Ready_Functions.R
\name{arrange.vars}
\alias{arrange.vars}
\title{arrange.vars}
\usage{
\method{arrange}{vars}(data, vars)
}
\arguments{
\item{data}{the name of your df}

\item{vars}{the names of your variables}
}
\value{
df
}
\description{
returns date time in a string seperated by underscores
}
\examples{
bckt_intl_bckt_frcsts <- arrange.vars(data = forecasts_2, vars = c("RGN" = 1,"P_STYL_CD" = 2,"PLNT_CD" = 3))
}
