% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Production_Ready_Functions.R
\name{better_summary}
\alias{better_summary}
\title{better_summary}
\usage{
better_summary(df)
}
\arguments{
\item{df}{your dataframe}
}
\description{
It's just a better summary
}
\examples{
mtcars \%>\% better_summary()
}
