% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Production_Ready_Functions.R
\name{cal_to_fisc_wk}
\alias{cal_to_fisc_wk}
\title{cal_to_fisc_wk}
\usage{
cal_to_fisc_wk()
}
\value{
"YYYY-mm-dd UTC"
}
\description{
This function returns the beginning of the current Steelcase fiscal week
}
