% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Production_Ready_Functions.R
\name{Chrctr_Crrnt_Dttme}
\alias{Chrctr_Crrnt_Dttme}
\title{Chrctr_Crrnt_Dttme}
\usage{
Chrctr_Crrnt_Dttme()
}
\value{
string
}
\description{
returns date time in a string seperated by underscores
}
\examples{
Chrctr_Crrnt_Dttme()
}
