% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Production_Ready_Functions.R
\name{get_fisc_wk}
\alias{get_fisc_wk}
\title{get_fisc_wk}
\usage{
get_fisc_wk()
}
\value{
integer 1-5 current fiscal week of period
}
\description{
This function returns current week of the fiscal period
}
