% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Production_Ready_Functions.R
\name{get_ver}
\alias{get_ver}
\title{get_ver}
\usage{
get_ver(week)
}
\arguments{
\item{week}{week of fiscal period}
}
\value{
0 or 1
}
\description{
This function returns whether or not it is the specified week of the fiscal period
}
