% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Production_Ready_Functions.R
\name{load_zip}
\alias{load_zip}
\title{load_zip}
\usage{
load_zip(the_file, extension)
}
\arguments{
\item{the_file}{zipped output file you want to load}

\item{extension}{the extension of the file once unzipped}
}
\description{
This function loads your output from a zip file
}
