% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Production_Ready_Functions.R
\name{per_error}
\alias{per_error}
\title{per_error}
\usage{
per_error(df, predict_var, actual_var)
}
\arguments{
\item{df}{the name of your df}

\item{predict_var}{the name of your prediction}

\item{actual_var}{the name of your actual}
}
\value{
dataframe
}
\description{
Computes percent error
}
\examples{
bckt_accrcy_02 <- per_error(df = bckt_accrcy_01, predict_var = predict, actual_var = SL_QT)
}
