% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Production_Ready_Functions.R
\name{rs_by_var_2}
\alias{rs_by_var_2}
\title{rs_by_var_2}
\usage{
rs_by_var_2(df, byvars, num)
}
\arguments{
\item{df}{the name of your dataframe}

\item{byvars}{the by variables you wish to sample by}

\item{num}{the number of samples to return}
}
\value{
subsetted dataframe
}
\description{
This function takes a random sample of your data based on inputted by variables.  Requires SQLDF
}
\examples{
rs_by_var_2(df = mtcars, byvars = c("cyl", "carb"), num = 2)
}
