% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Production_Ready_Functions.R
\name{save_time_stamped}
\alias{save_time_stamped}
\title{save_time_stamped}
\usage{
save_time_stamped(df, the_file_path, the_time_stamp, row.names = F)
}
\arguments{
\item{df}{the name of your dataframe}

\item{the_file_path}{out file path}

\item{the_time_stamp}{value to timestamp your output with}
}
\description{
This function saves a time stamped version of your output data set, in addtion to the data set itself
}
