% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Production_Ready_Functions.R
\name{save_zip}
\alias{save_zip}
\title{save_zip}
\usage{
save_zip(the_file)
}
\arguments{
\item{the_file}{output file you want to zip}
}
\description{
This function saves your output file as a zip file
}
