% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Production_Ready_Functions.R
\name{arrange.vars}
\alias{arrange.vars}
\title{vars' must be a named vector, e.g. c("var.name"=1)}
\usage{
arrange.vars(data, vars)
}
\description{
vars' must be a named vector, e.g. c("var.name"=1)
}
